Move out some Elasticsearch-based code into a common lib
