+++ common-lib/src/main/scala/uk/ac/wellcome/finatra/exceptions/ElasticsearchExceptionMapper.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.api.exceptions
+package uk.ac.wellcome.finatra.exceptions
 
 import org.elasticsearch.ElasticsearchException
 import com.twitter.finagle.http.{Request, Response}
