+++ common-lib/src/main/scala/uk/ac/wellcome/finatra/services/ElasticsearchService.scala
@@ -1,8 +1,9 @@
-package uk.ac.wellcome.platform.api.services
+package uk.ac.wellcome.finatra.services
 
 import javax.inject.{Inject, Singleton}
-import uk.ac.wellcome.platform.api.modules.ElasticClientModule
 import com.sksamuel.elastic4s.ElasticClient
+
+import uk.ac.wellcome.finatra.modules.ElasticClientModule
 
 
 @Singleton
